requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
alembic
redis
pydantic[email]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
email-validator
python-multipart
pytest
pytest-asyncio
httpx
pytest-mock
coverage
aiosqlite
locust